Rework updater implementation to split recovery/normal builds
